requirements.txt
pytest~=7.4.2
money~=1.3.0
holidays~=0.33
scipy~=1.11.2
pandas~=2.1.1
businessdate~=0.6
pendulum~=2.1.2
financepy~=0.310
python-dateutil~=2.8.2
QuantLib~=1.31.1